Read Counting

Count reads by gene and by exon

count_reads(bam = NULL, ext = ".bam", by = c("gene", "exon"),
  samples.annotation = NULL, result.dir = "COUNT", save = TRUE,
  gtf = "/eqmoreaux/genomes/Homo_sapiens/Ensembl/GRCh37/Annotation/Genes/genes.gtf",
  pairedEnd = TRUE, ignore.strand = FALSE, count.mode = "Union",
  thread = 10)

Arguments

bam

a vector of bam files (in the same directory) or a directory containing one or more paths to bam files. if NULL, BAM files in the current working directory are choosed.

ext

BAM file extension. Default is ".bam". Used only when bam is a directory

by

One of "gene" (for counting in gene) or "exon" (for counting in exon)

samples.annotation

the path to "samples.txt" tabulation file containing sample annotations. This file should contain at least 3 columns:

name: corresponding to sample name. It will be used to report the result

fastq1: the first fastq file name for paired-end sequencing

fastq2: the second fastq file name for paired-end sequencing. This is optional if single-end sequencing is used

result.dir

a directory to save the output.

save

logical value; if TRUE, the result is saved in result.dir.

gtf

gene annotation file.

pairedEnd

specify if the data are paired-end sequencing data. Default is TRUE.

ignore.strand

A logical indicating if strand should be considered when matching. if TRUE, allows minus strand reads to count in plus strand genes, and vice versa. Should be FALSE for stranded RNAsequencing.

count.mode

counting mode. Read ?GenomicAlignments::summarizeOverlaps. Default value is "Union". Reads that overlap any portion of exactly one feature are counted. Reads that overlap multiple features are discarded.

thread

Number of threads to be used. This depends to the available computer ressources.

Value

Return an object of class SummarizedExperiment

Details

Used Bioconductor packages:

GenomicFeatures to prepare the transcript database

Rsamtools to import SAM/BAM files in R

GenomicAlignments for read counting

BiocParallelfor parallel computing

Input files: BAM files. No need to sort the BAM files in Bioconductor version > 2.12.