Generate Genome Index

Create reference genome index for STAR software

create_star_index(fasta_file, gtf_file, thread = 4, starIndex_dir = NULL)

Arguments

fasta_file

path to reference genome fasta file. If the genome has been downloaded from http://ccb.jhu.edu/software/tophat/igenomes.shtml, the path to the fasta file is of form: "~/genomes/Homo_sapiens/UCSC/hg19/Sequence/WholeGenomeFasta/genome.fa"

gtf_file

path to annotation gtf file. If iGenomes, the path can be of form: "~/genomes/Homo_sapiens/UCSC/hg19/Annotation/Genes/genes.gtf"

thread

the number of threads to be used. Dpending on the available ressources.

starIndex_dir

a directory to hold STAR index files. If NULL a subdirectory named "StarIndex" is created the same directory as the fasta file.

Examples

create_star_index(fasta_file = "~/genomes/sequence/genome.fa",
   gtf_file = "~/genomes/annotation/genes.gtf")

#> STAR index has been created at  /Users/kassambara/genomes/sequence/StarIndex 
#>