Expressed genes in each cell population

Identification of expressed genes in each cell population. Expressed genes in a given cell population is defined as genes with average expression above a given cutoff.

expressed(x, groups, cutoff = 1)

# S3 method for expressed
summary(object, ...)

Arguments

x

matrix or numeric data frame containing gene expression data. Rows are genes and columns are samples.

groups

a factor containing the group assignement of samples

cutoff

a numeric value specifying the threshold above which a gene is declared as expressed.

object

an object of class expressed

...

additional arguments.

Value

expressed() returns an object of class "expressed". Columns are samples or groups and rows are genes. Values can be 1 (if expressed) or 0 (if not expressed).

summarize() returns a list containing the following components:

exprs_in_all: The number of expressed genes in all groups

exprs_in_one: The number of expressed genes in at least one group

exprs_in_each: a vector containing the number of genes expressed in each group

exprs_in_specific: the number of genes expressed only in specific group