Differential Expression Between Groups

Identify differentially expressed genes between groups:

get_diff_1(): Common up/down regulated genes in one group compared to each of the other groups in pairwise analysis. It's an alias of get_diff_pwc() when length(grp1) = 1.

get_diff_2(): Differentially expressed genes between two groups in a simple two class analysis. Alias of get_diff_pwc() when length(grp1) = length(grp2) = 1.

get_diff_n(): Differentially expressed genes between multiple groups using pairewise analysis or not (see details section).

get_diff_pwc(): Pairwise comparisons (pwc) between sample classes in grp1 (e.g.: grp1 = c("A", "B", ...)) and sample classes in grp2 (e.g.: grp2 = c("C", "D", ...) ). Can do all the above mentioned type of analyses. It's an alias of get_diff_n(), get_diff_1( ) and get_diff_2().

significant(): Take an object of class "get_diff", "lm_results" or "DESeqResult" and returns significantly differentially expressed genes with a given cutoff for false discovery rate and fold change.

For more information about these functions, read details section.

get_diff_1(dds, factor, grp1, grp2 = NULL, alpha = 0.05, fc = 1.5,
  active_exprs = 1, count.norm = NULL, exprs.norm = NULL,
  verbose = TRUE, ...)

get_diff_2(dds, factor, grp1, grp2, alpha = 0.05, fc = 1.5,
  active_exprs = 1, count.norm = NULL, exprs.norm = NULL,
  verbose = TRUE, ...)

get_diff_n(dds, factor, grp1, grp2 = NULL, alpha = 0.05, fc = 1.5,
  active_exprs = 1, count.norm = NULL, exprs.norm = NULL, merge = FALSE,
  verbose = TRUE, ...)

get_diff_pwc(dds, factor, grp1, grp2 = NULL, alpha = 0.05, fc = 1.5,
  active_exprs = 1, count.norm = NULL, exprs.norm = NULL, merge = FALSE,
  verbose = TRUE, ...)

significant(object, fdr = 0.05, fc = 1.5, orderby = c("padj", "lfc"))

Arguments

dds

an object of class:

DESeqDataSet containing the results of the function DESeq(), or

lm_fit

dds must contain the information about samples.

factor

the name of a factor in the design formula.

grp1, grp2

a list of 2 character vectors containing the groups to be compared. The argument grp1 is used as numerator and grp2 as the denominator for computing the fold change (See details section). If grp2 is NULL, then grp1 is compared to all the other groups in the data.

alpha

The significance cutoff used for optimizing the independent filtering (by default 0.05); see ?DESeq2::results.

fc

the fold change threshold. Only genes with an absolute fold change >= fc and adjusted p-value <= fdr and a detection_call = 1, are considered as significantly differentially expressed.

active_exprs

a numeric value specifying detected expression cutoff. If not NULL, any gene with an average expression above this cutoff, in at least one group, is declared as actively expressed. Used for defining presence and absence detection call. If not NULL, a column named detection_call is added to the result. Possible values for detection_call is 0 (the gene is not actively expressed) or 1 (the gene is actively expressed).

count.norm

a data frame containing the normalized count as provided by DESeq2. Only used for computing the detection call when active_expression is not NULL.

exprs.norm

a data frame containing the normalized gene expression data. An alias of count.norm.

verbose

logical value; if TRUE the progression is shown. Default is TRUE.

...

other arguments to be passed to the functions DESeq2:results() or lm_results()

merge

a logical value. Default is FALSE. Used only in multiclass comparison cases (seee details section).

object

an object of class get_diff, DE_Results, DESeqResults or a data frame containing the columns padj, log2FoldChange and optionally the column detection_call.

fdr

Accepted false discovery rate for considering genes as differentially expressed.

orderby

possible values are "padj" and "lfc" for ordering significant genes by the adjusted p-value or the log2FoldChange

Value

The functions get_diff_*() returns an object of class DESeqResults (see ?DESeq2::results) and "get_diff" which contains the following columns: baseMean, log2FoldChange, lfcSE, stat, pvalue, padj, detection_call and significance. To see the description of the columns use the function mcols(res, use.names = TRUE), where res is the result of get_diff_*(). The function plot() can be used to draw the results. (See ?plot_MA).

significant() returns an object of class "significant" containing significantly differentially expressed genes with a given cutoff for false discovery rate (e.g.: 0.05) and fold change (e.g.: 1.5) and, a detection_call =1 (if available in the argument object). The object "significant" is a list with the following components:

up: data frame containing upregulated genes in grp1 compared to grp2

down: data frame containing downregulated genes in grp1 compared to grp2

nup: number of upregulated genes

ndown: number of downregulated genes

Details

Definition of general terms contained in the results returned by get_diff_*() functions:

The log2FoldChange is calculated using grp1 as numerator and grp2 as denominator.

The baseMean is calculated as mean of normalized counts for all samples.

The detection_call defines gene presence/absence detection calls which possible values are: 0 (the gene is not actively expressed) or 1 (the gene is actively expressed, i.e. normalize expression >= active_exprs cutoff).

significance: Deregulation significance: Up = 1; Down = -1; Not significant = 0. A gene is significantly deregulated if a) the detection_call = 1, b) the fold change >= fc cutoff and c) the padj < = alpha cutoff.

Possible values for the arguments grp1 and grp2 can be of the forms:

grp1 = "A" and grp2 = "B": Simple two class comparisons. Recommended functions: get_diff_2() or get_diff_pwc(). The adjusted p-values (padj) and the log2 fold changes (log2FoldChange) are those returned by DESeq2::results.

grp1 = "A" and grp2 = c("B", "C", ...). One class compared to more than one other groups. Recommended functions: get_diff_1() or get_diff_pwc() for identifying common up/down regulated genes in group A compared to each of groups B and C using pairwise analysis (e.g.: A vs B and A vs C).

Adjusted p-value: The maximum of all the adjusted p-values (padj) obtained from the different pairwise comparisons is used as the final padj in the results. For a given gene, a max(padj) < 0.05 indicates that each of the padj is < 0.05 in all pairwise comparisons. If a gene is upregulated in one comparison but downregulated in another comparison, then the max(padj) is set to NA. Only genes commonly up- or down- regulated in all the comparisons are considered to be specific gene signatures of grp1.

log2 fold change: The minimum of all log2FoldChange (lfc) obtained from the different pairwise comparisons is used as the final log2FoldChange. For a given gene, min(lfc) > 1.5 in absolute value (abs) indicates that each of the abs(lfcs) is > 1.5 in all pairwise comparisons. If a gene is upregulated in one comparison but downregulated in another comparison, then the min(lfc) is set to 0. Only genes commonly up- or down- regulated in all the comparisons are considered to be specific gene signatures of grp1.

lfcSE, stat and pvalue: These columns can be ignored in the result as they are obtained by performing a global comparison (not a pairwise comparison) between grp1 and grp2. We kept them only for conveniance with DESeq2 result formats.

grp1 = c("A", "B", ...) and grp2 = c("D", "E", ...); Multiple class comparisons. Recommended functions: get_diff_n() or get_diff_pwc().

If merge = TRUE, then A+B are merged and compared to merged D+E in a single two class analysis. The adjusted p-values (padj) and the log2 fold changes (log2FoldChange) are those returned by DESeq2::results.

If merge = FALSE, then separated pairewise comparisons are performed using get_diff_pwc(), for example as follow: "A" vs c("D", "E", ...) and "B" vs c("D", "E", ...). In this particular case, the different comparisons are A vs D, A vs E, B vs D and B vs E, etc. Kept genes in the final results, are only those commonly up/down regulated in both A and B compared to D and E. The returned results have the same format as those returned by get_diff_1(). Read the paragraphs above concerning adjusted p-value and log2 fold change.