Identification of a transcription regulators in a list of genes

Identification of DNA-binding transcription factors and/or epigenetic enzymes in a list of genes. Only data for human are available (see ?human.tf and ?human.ee).

get_human_regulators(ids, id_type = "ensembl", keep.all = TRUE)

Arguments

ids

a vector containing gene Ensembl Ids ("NSG00000130816") or gene names ("DNMT1"). if ids is a matrix or a data.frame, then the row names are used as ids.

id_type

the type of the ids. Possible values are "ensembl" or "name".

keep.all

a logical value. If TRUE, all the genes are kept in the final result. If FALSE, only transcription factors/epigenetic enzymes are returned

Value

Returns an object of class data.frame containing the following columns:

id: gene ids

TF: with a value of 1 (the genes is transcription factor) or 0 (the gene is not a transcription factor).

EE_class: specifies the epigenetic enzyme class of the genes.