Human epigenetic enzymes

A list of 212 epigenetic enzymes, including writers, readers, editors and erasers of the most important epigenetic marks.

data(human.ee)

Format

A data frame with 4 columns and 212 rows coorresponding to genes.

Source

Yang Z, Jones A, Widschwendter M, Teschendorff AE. An integrative pan-cancer-wide analysis of epigenetic enzymes reveals universal patterns of epigenomic deregulation in cancer. Genome Biol. 2015 Jul 14;16:140. doi: 10.1186/s13059-015-0699-9.

Examples

 data(human.ee)
 head(human.ee)

#>                   name            group                    class
#> ENSG00000130816  DNMT1 DNA modification DNA methylation maintain
#> ENSG00000119772 DNMT3A DNA modification      De novo methylation
#> ENSG00000088305 DNMT3B DNA modification      De novo methylation
#> ENSG00000142182 DNMT3L DNA modification      De novo methylation
#> ENSG00000141644   MBD1 DNA modification   DNA methylation reader
#> ENSG00000134046   MBD2 DNA modification   DNA methylation reader
#>                         ensembl
#> ENSG00000130816 ENSG00000130816
#> ENSG00000119772 ENSG00000119772
#> ENSG00000088305 ENSG00000088305
#> ENSG00000142182 ENSG00000142182
#> ENSG00000141644 ENSG00000141644
#> ENSG00000134046 ENSG00000134046