MA-plot from means and log fold changes

Make MA-plot which is a scatter plot of log2 fold changes (on the y-axis) versus the mean expression signal (on the x-axis).

plot_MA(x, fdr = 0.05, fc = 1.5, detection_call = NULL, ylim = NULL,
  colNonSig = "darkgray", colUp = "#BC002A", colDown = "#16614E",
  colLine = "black", log = "x", cex = 0.45, xlab = "Mean expression",
  ylab = "Log2 fold change", addlegend = TRUE, highlight = 10, ...)

# S3 method for DESeqResults
plot(x, ...)

# S3 method for DE_Results
plot(x, ...)

Arguments

x

an object of class DESeqResults, get_diff, DE_Results, matrix or data frame containing the columns baseMean, log2FoldChange, and padj. Rows are genes.

baseMean: the mean expression of genes in the two groups.

log2FoldChange: the log2 fold changes of group 2 compared to group 1

padj: the adjusted p-value of the used statiscal test.

fdr

Accepted false discovery rate for considering genes as differentially expressed.

fc

the fold change threshold. Only genes with a fold change >= fc and padj <= fdr are considered as significantly differentially expressed.

detection_call

a numeric vector with length = nrow(x), specifying if the genes is expressed or not. Default is NULL. If detection_call column is available in x, it will be used.

ylim

The limits for the y-axis. If missing, an attempt is made to choose a sensible value. Dots exceeding the limits will be displayed as triangles at the limits, pointing outwards.

colNonSig, colUp, colDown

colors to be used for non-significant, up-regulated and down-regulated data points, respectively.

colLine

colour to use for the horizontal (y=0) line.

log

which axis/axes should be logarithmic; will be passed to plot.

cex

point size

xlab, ylab

x and y axis labels

addlegend

a logical value. If TRUE a legend is added to the plot.

highlight

highlight top genes. Default value is 10 for top 10 genes.

...

other parameters to be passed to the function plot

Value

returns a base plot