Read Aligning

Read aligning using STAR software.

star_align(data_dir = getwd(), samples.annotation = "samples.txt",
  star.index = "/eqmoreaux/genomes/Homo_sapiens/Ensembl/GRCh37/Sequence/StarIndex/",
  result.dir = getwd(), keep = c("name_sorted_bam"), pairedEnd = TRUE,
  fastq.gz = TRUE, thread = 10)

Arguments

data_dir

the directory containing the data. It should contains a subdirectory called "FASTQ" which contains fastq files. Default value is the working directory

samples.annotation

the path to "samples.txt" tabulation file containing sample annotations. This file should contain at least 3 columns:

name: corresponding to sample name. It will be used to report the result

fastq1: the first fastq file name for paired-end sequencing

fastq2: the second fastq file name for paired-end sequencing. This is optional if single-end sequencing is used

star.index

reference genome for STAR.

result.dir

The absolute path of result directory.

keep

a character vector specifying the file to be kept. Allowed values include c("name_sorted_bam", "chr_sorted_bam", "sam" ).

pairedEnd

specify if the data are paired-end sequencing data. Default is TRUE.

fastq.gz

specify if the FASTQ files are compressed. Default is TRUE.

thread

Number of threads to be used. This depends to the available computer ressources.

Value

Three subdirectories are created:

BAM: containing unsorted and sorted BAM files, and BAM index file. BAM comtent is sorted by read names (*_name_sorted.bam) or by chromosome (*_sorted.bam). The index files are of form *_sorted.bam.bai. *_name_sorted.bam files are used for read counting using hseq-count or R. *_sorted.bam and *_sorted.bam.bai files can be used for IGV visualization.

COUNT: containing read counting results. It contains the following files: 1. se.RDATA: containing an object "se" which is an object of class SummarizedExperiment

Details

The function star_align(), requires STAR and samtools programs to work. Make sure that they are installed. The workflow is as follow: 1. Align all FASTQ files using STAR. BAM files are generated. 2. Organize BAM files (sorting and indexing). Samtools program required. 3. Count read using Bioconductor packages: "GenomicFeatures", "Rsamtools", "GenomicAlignments" and "BiocParallel" required.